requirements.txt
cryptography==2.7
Flask==1.1.1
Flask-MySQL==1.5.2
Flask-MySQLdb==0.2.0
gunicorn==20.1.0
mysql-connector-python==8.0.18
mysqlclient==2.0.3
numpy==1.20.2
Padding==0.5
pycrypto==2.6.1
scipy==1.3.1
SQLAlchemy==1.3.9
Werkzeug==0.16.0